Rename quoted.hpp to wrappers.hpp
