+++ include/glaze/json.hpp
@@ -12,8 +12,8 @@
 #include "glaze/json/ndjson.hpp"
 #include "glaze/json/prettify.hpp"
 #include "glaze/json/ptr.hpp"
-#include "glaze/json/quoted.hpp"
 #include "glaze/json/raw_string.hpp"
 #include "glaze/json/read.hpp"
 #include "glaze/json/schema.hpp"
+#include "glaze/json/wrappers.hpp"
 #include "glaze/json/write.hpp"
