+++ include/glaze/json/schema.hpp
@@ -18,7 +18,7 @@
 #include "glaze/api/std/vector.hpp"
 #include "glaze/api/tuplet.hpp"
 #include "glaze/api/type_support.hpp"
-#include "glaze/json/quoted.hpp"
+#include "glaze/json/wrappers.hpp"
 #include "glaze/json/write.hpp"
 
 namespace glz
